PO-8928 : refactor - getPaymentTerms
