+++ src/main/java/uk/gov/hmcts/opal/service/iface/DefendantAccountPaymentTermsServiceInterface.java
@@ -2,6 +2,7 @@ package uk.gov.hmcts.opal.service.iface;
 
 import uk.gov.hmcts.opal.dto.AddPaymentCardRequestResponse;
 import uk.gov.hmcts.opal.dto.GetDefendantAccountPaymentTermsResponse;
+import uk.gov.hmcts.opal.dto.request.AddDefendantAccountPaymentTermsRequest;
 
 public interface DefendantAccountPaymentTermsServiceInterface {
 
@@ -10,4 +11,11 @@ public interface DefendantAccountPaymentTermsServiceInterface {
     AddPaymentCardRequestResponse addPaymentCardRequest(Long defendantAccountId, String businessUnitId,
         String businessUnitUserId,
         String postedByName, String ifMatch);
+
+    GetDefendantAccountPaymentTermsResponse addPaymentTerms(Long defendantAccountId,
+        String businessUnitId,
+        String businessUnitUserId,
+        String postedByName,
+        String ifMatch,
+        AddDefendantAccountPaymentTermsRequest addPaymentTermsRequest);
 }
